requirements.txt
undetected-chromedriver
python-telegram-bot>=20
